cleaned up the move character function. Created isValidDirection function and isLocal function to make code cleaner and easier to read. The isLocal function will need to be changed based on how we represent non-local rooms.
Also rearranged some of the order of things in the non-local part of
move character. The character needs the RoomIn field updated BEFORE it
is sent to the new world, otherwise it will not work correctly. 

Also replaced the message creation stuff with the helper functions from
the FormattedString class.