I implemented an alternate trading mechanism using channels. It seems to work rather well so far. Part of the reason it came out so well was because I rewrote some of the inventory functions. The method we were using, of keeping an items actual structure only once with an int for the quantity was incorrect. Doing it this way meant whenever we returned that item we returned a pointer to the exact same item.
I think whith some of the changes I made to the
Character and Inventory class, this version of trading is very clean an
easy to read; and it seems to test well.

I will hold off merging it into the master branch until monday incase
you wanted to discuss it.